Convert endowment hardening tests
