+++ packages/snaps-execution-environments/src/common/endowments/endowments.test.browser.ts
@@ -1,0 +1,315 @@
+/* eslint-disable import/no-unassigned-import, @typescript-eslint/restrict-template-expressions */
+
+import 'ses';
+
+import { walkAndSearch } from '../test-utils/endowments';
+import { testEndowmentHardening } from '../test-utils/hardening';
+import buildCommonEndowments from './commonEndowmentFactory';
+import Crypto from './crypto';
+import date from './date';
+import interval from './interval';
+import math from './math';
+import network from './network';
+import timeout from './timeout';
+
+// @ts-expect-error - `globalThis.process` is not optional.
+delete globalThis.process;
+
+lockdown({
+  domainTaming: 'unsafe',
+  errorTaming: 'unsafe',
+  stackFiltering: 'verbose',
+});
+
+describe('endowments', () => {
+  describe('hardening', () => {
+    const modules = buildCommonEndowments();
+    modules.forEach((endowment) => endowment.factory());
+
+    // Specially attenuated endowments or endowments that require
+    // to be imported in a different way
+    const { SubtleCrypto, crypto } = Crypto.factory();
+    const {
+      setTimeout: setTimeoutAttenuated,
+      clearTimeout: clearTimeoutAttenuated,
+    } = timeout.factory();
+    const {
+      setInterval: setIntervalAttenuated,
+      clearInterval: clearIntervalAttenuated,
+    } = interval.factory();
+    const { Math: mathAttenuated } = math.factory();
+    const { fetch: fetchAttenuated } = network.factory();
+    const { Date: DateAttenuated } = date.factory();
+
+    const TEST_ENDOWMENTS = {
+      // Constructor functions.
+      BigInt: {
+        endowments: { BigInt },
+        factory: () => BigInt(3),
+      },
+      SubtleCrypto: {
+        endowments: { SubtleCrypto },
+        factory: () => undefined,
+      },
+      TextDecoder: {
+        endowments: { TextDecoder },
+        factory: () => new TextDecoder(),
+      },
+      TextEncoder: {
+        endowments: { TextEncoder },
+        factory: () => new TextEncoder(),
+      },
+      URL: {
+        endowments: { URL },
+        factory: () => new URL('https://metamask.io/snaps/'),
+      },
+      Int8Array: {
+        endowments: { Int8Array },
+        factory: () => new Int8Array(),
+      },
+      Uint8Array: {
+        endowments: { Uint8Array },
+        factory: () => new Uint8Array(),
+      },
+      Uint8ClampedArray: {
+        endowments: { Uint8ClampedArray },
+        factory: () => new Uint8ClampedArray(),
+      },
+      Int16Array: {
+        endowments: { Int16Array },
+        factory: () => new Int16Array(),
+      },
+      Uint16Array: {
+        endowments: { Uint16Array },
+        factory: () => new Uint16Array(),
+      },
+      Int32Array: {
+        endowments: { Int32Array },
+        factory: () => new Int32Array(),
+      },
+      Uint32Array: {
+        endowments: { Uint32Array },
+        factory: () => new Uint32Array(),
+      },
+      Float32Array: {
+        endowments: { Float32Array },
+        factory: () => new Float32Array(),
+      },
+      Float64Array: {
+        endowments: { Float64Array },
+        factory: () => new Float64Array(),
+      },
+      BigInt64Array: {
+        endowments: { BigInt64Array },
+        factory: () => new BigInt64Array(),
+      },
+      BigUint64Array: {
+        endowments: { BigUint64Array },
+        factory: () => new BigUint64Array(),
+      },
+      DataView: {
+        endowments: { DataView, ArrayBuffer },
+        factory: () => new DataView(new ArrayBuffer(64)),
+      },
+      ArrayBuffer: {
+        endowments: { ArrayBuffer },
+        factory: () => new ArrayBuffer(64),
+      },
+      AbortController: {
+        endowments: { AbortController },
+        factory: () => new AbortController(),
+      },
+      AbortSignal: {
+        endowments: { AbortSignal },
+        // @ts-expect-error - `abort()` method exists in browser, but not in
+        // Node.js.
+        factory: () => AbortSignal.abort(),
+      },
+      DateAttenuated: {
+        endowments: { DateAttenuated },
+        factory: () => new DateAttenuated(),
+      },
+
+      // Objects.
+      console: {
+        endowments: { console },
+        factory: () => console,
+      },
+      crypto: {
+        endowments: { crypto },
+        factory: () => crypto,
+      },
+      mathAttenuated: {
+        endowments: { mathAttenuated },
+        factory: () => mathAttenuated,
+      },
+      WebAssembly: {
+        endowments: { WebAssembly },
+        factory: () => WebAssembly,
+      },
+
+      // Functions.
+      atob: {
+        endowments: { atob },
+        factory: () => atob('U25hcHM='),
+      },
+      btoa: {
+        endowments: { btoa },
+        factory: () => btoa('Snaps'),
+      },
+      setTimeoutAttenuated: {
+        endowments: { setTimeoutAttenuated },
+        factory: () => setTimeoutAttenuated((param: unknown) => param, 1),
+      },
+      clearTimeoutAttenuated: {
+        endowments: { clearTimeoutAttenuated },
+        factory: () => undefined,
+      },
+      setIntervalAttenuated: {
+        endowments: { setIntervalAttenuated },
+        factory: () => setIntervalAttenuated((param: unknown) => param, 100000),
+      },
+      clearIntervalAttenuated: {
+        endowments: { clearIntervalAttenuated },
+        factory: () => undefined,
+      },
+      fetchAttenuated: {
+        endowments: { fetchAttenuated },
+        factory: () => undefined,
+      },
+    };
+
+    Object.entries(TEST_ENDOWMENTS).forEach(
+      ([name, { endowments, factory }]) => {
+        it(`hardens ${name}`, () => {
+          const sourceCode = `;(${testEndowmentHardening})(${name},${factory})`;
+          const subject = endowments[name as keyof typeof endowments] as {
+            // eslint-disable-next-line @typescript-eslint/naming-convention
+            __flag?: unknown;
+          };
+
+          const compartment = new Compartment(endowments, {}, {});
+
+          const errors = compartment.evaluate(sourceCode);
+          const instance = factory();
+
+          expect(subject.__flag).toBeUndefined();
+          expect(instance?.__flag).toBeUndefined();
+          expect(errors.length).toBeGreaterThan(0);
+        });
+
+        if (factory()) {
+          it('does not leak `globalThis`', () => {
+            const instance = factory();
+            expect(walkAndSearch(instance, globalThis)).toBe(false);
+          });
+        }
+      },
+    );
+  });
+
+  describe('commonEndowmentFactory', () => {
+    it('creates endowments with the expected properties', () => {
+      const endowments = buildCommonEndowments();
+      expect(endowments).toStrictEqual([
+        {
+          factory: expect.any(Function),
+          names: ['crypto', 'SubtleCrypto'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['setInterval', 'clearInterval'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Math'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['fetch'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['setTimeout', 'clearTimeout'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['TextDecoder'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['TextEncoder'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Date'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['AbortController'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['AbortSignal'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['ArrayBuffer'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['atob'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['BigInt'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['BigInt64Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['BigUint64Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['btoa'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['DataView'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Float32Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Float64Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Int8Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Int16Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Int32Array'],
+        },
+        {
+          factory: expect.any(Function),
+          names: ['Uint8Array'],
+        },
+        { factory: expect.any(Function), names: ['Uint8ClampedArray'] },
+        { factory: expect.any(Function), names: ['Uint16Array'] },
+        { factory: expect.any(Function), names: ['Uint32Array'] },
+        { factory: expect.any(Function), names: ['URL'] },
+        { factory: expect.any(Function), names: ['WebAssembly'] },
+      ]);
+    });
+  });
+});
